PEP8: Renames Scripts package
