+++ tmt_future/scripts/Sync_Photometry_Video.py
